requirements.txt
streamlit
pandas
requests
python-dotenv
autogen-agentchat
autogen-ext
newsapi-python
openai
tiktoken
duckdb
scikit-learn